pyproject.toml
[tool.poetry]
name = "research-agent"
version = "0.1.0"
description = "AG2 Scientific Literature Review Platform - A prototype system for automated literature review using AG2 agents"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "demo"}]

[tool.poetry.dependencies]
python = ">=3.11, <3.14"
pyautogen = ">=0.2.30"
arxiv = ">=2.1.0"
requests = ">=2.31.0"
openai = ">=1.40.0"
scikit-learn = ">=1.3.0"
numpy = ">=1.24.0"
nltk = ">=3.8.0"
pandas = ">=2.0.0"
beautifulsoup4 = ">=4.12.0"
click = ">=8.1.0"
python-dotenv = ">=1.0.0"
aiohttp = ">=3.9.0"
ag2 = "^0.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.0.0"
flake8 = "^6.0.0"

[tool.poetry.scripts]
research-agent = "demo.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3